Initial commit
This commit basically adds a verbatim copy of the existing NixOS
configuration.

The repo is structured similarly to Stormhub for now (might even get
merged at some point), but the idea is:
- Use standardized Nix formatters (nixfmt)
- Setup Agenix to not commit the SMB share password used in the original
  configuration
- Keep _some_ structure for now

This will be adjusted in the future. For now, I want to flake my config
so it can be version controlled for once. Ideally, with time, we move to
Home Manager